Adds required folder name in README.md
The index.js requires the plugin to be located in a specific folder called "trainTeleports"
instead of using a relative or dynamic directory. When cloning the repository directly into
sharedPlugins it doesn't work until the folder has been renamed correctly.

This change will clarify the correct installation path of the plugin.